Add test case
